refactor: página de minhasCompras
