+++ src/main/webapp/css/paginas/minhasCompras.css
@@ -1,6 +1,3 @@
-
-@import url(/css/style.css);
-
 #container-compras {
     min-height: 280px;
 }
