+++ src/main/webapp/js/componentes/pedido/ContainerItens.js
@@ -1,4 +1,4 @@
-import { formatarPreco } from "../../../../../js/script.js";
+import { formatarPreco } from "../../script.js";
 
 export class ContainerItens extends HTMLDivElement {
     constructor(itensPedido){
