--- src/main/webapp/paginas/cliente/minhasCompras/minhasCompras.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-BR">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Minhas Compras</title>
-    <link rel="stylesheet" href="/paginas/cliente/minhasCompras/css/minhasCompras.css">
-</head>
-
-<body class="bg-light">
-    <nav id="main-nav"></nav>
-
-    <div class="container my-5">
-        <h1 class="mb-4 text-center">Minhas Compras</h1>
-
-        <div id="container-compras" class="container mt-4"></div>
-    </div>
-
-    <footer id="main-footer"></footer>
-
-    <script type="module" src="js/loadPage.js"></script>
-    <script type="module" src="/paginas/cliente/minhasCompras/js/minhasCompras.js"></script>
-</body>
-
-</html>
