+++ src/main/webapp/paginas/minhasCompras.html
@@ -1,0 +1,11 @@
+<title>Minhas Compras</title>
+
+<nav id="main-nav"></nav>
+
+<div class="container my-5">
+    <h1 class="mb-4 text-center">Minhas Compras</h1>
+
+    <div id="container-compras" class="container mt-4"></div>
+</div>
+
+<footer id="main-footer"></footer>
